# copiloto-function/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Copiloto Function Gateway",
    "version": "3.5",
    "description": "Sistema completo con memoria automática integrada, queries dinámicas, consulta previa y guardado automático de interacciones en Cosmos DB. Nota: los endpoints legacy /api/probar-endpoint e /api/invocar fueron retirados; usa /api/copiloto como router semántico.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://copiloto-func.ngrok.app"
    },
    {
      "url": "https://copiloto-semantico-func-us2.azurewebsites.net"
    }
  ],
  "paths": {
...
    "/api/diagnostico-recursos": {
      "get": {
        "summary": "Diagnóstico completo de recursos Azure",
        "operationId": "diagnosticoRecursos",
        "tags": [
          "Diagnóstico"
        ],
        "security": [],
        "parameters": [
          {
            "name": "metricas",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "costos",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "recurso",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diagnóstico completo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string"
                    },
                    "ambiente": {
                      "type": "string"
                    },
                    "recursos": {
                      "type": "object"
                    },
                    "metricas": {
                      "type": "object"
                    },
                    "alertas": {
                      "type": "array"
                    },
                    "recomendaciones": {
                      "type": "array"
                    },
                    "sistema": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          }
        }
      },
      "post": {
        "summary": "Diagnóstico con parámetros específicos",
        "operationId": "diagnosticoRecursosPost",
        "tags": [
          "Diagnóstico"
        ],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metricas": {
                    "type": "boolean"
                  },
                  "costos": {
                    "type": "boolean"
                  },
                  "recurso": {
                    "type": "string"
                  },
                  "profundidad": {
                    "type": "string",
                    "enum": [
                      "basico",
                      "detallado",
                      "completo"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Diagnóstico ejecutado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          }
        }
      }
    },
...
    "/api/copiloto": {
      "get": {
        "summary": "🧠 COPILOTO CON MEMORIA Y QUERIES DINÁMICAS - Endpoint principal con historial automático",
        "description": "Endpoint principal del copiloto que incluye consulta automática de memoria, queries dinámicas para filtrado avanzado, detección de intenciones de historial y respuesta directa con contexto de conversación. Integra automáticamente el sistema de memoria de Cosmos DB. Soporta filtros: tipo, contiene, endpoint, exito, fecha_inicio, fecha_fin, orden, limite.",
        "operationId": "copiloto",
        "tags": [
          "🧠 Memoria Integrada"
        ],
        "security": [],
        "parameters": [
          {
            "name": "Session-ID",
            "in": "header",
            "description": "ID de sesión para continuidad de conversación (prioridad alta)",
            "schema": {
              "type": "string",
              "example": "test_deduplicado_001"
            }
          },
          {
            "name": "Agent-ID",
            "in": "header",
            "description": "ID del agente para contexto de memoria (prioridad alta)",
            "schema": {
              "type": "string",
              "example": "TestAgent"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "description": "ID de sesión como parámetro de consulta",
            "schema": {
              "type": "string",
              "example": "test_deduplicado_001"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "description": "ID del agente como parámetro de consulta",
            "schema": {
              "type": "string",
              "example": "TestAgent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta del copiloto con memoria integrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "copiloto": {
                      "type": "string"
                    },
                    "memoria_activa": {
                      "type": "boolean"
                    },
                    "contexto_conversacion": {
                      "type": "object",
                      "properties": {
                        "mensaje": {
                          "type": "string"
                        },
                        "ultimas_consultas": {
                          "type": "string"
                        },
                        "session_id": {
                          "type": "string"
                        }
                      }
                    },
                    "historial_detectado": {
                      "type": "boolean"
                    },
                    "interacciones_previas": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/diagnostico": {
      "get": {
        "summary": "🔍 DIAGNÓSTICO DE SESIÓN - Analiza qué ocurrió en una sesión",
        "description": "Analiza interacciones de una sesión específica, detecta patrones de errores y genera diagnóstico con métricas.",
        "operationId": "diagnostico",
        "tags": [
          "🧠 Memoria Integrada"
        ],
        "security": [],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "description": "ID de sesión a diagnosticar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diagnóstico completado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exito": {
                      "type": "boolean"
                    },
                    "diagnostico": {
                      "type": "object"
                    },
                    "metricas": {
                      "type": "object"
                    },
                    "patrones_detectados": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}